Move tape extension and filechooser to Tape
